07-streaming/src/job/homework_5min_tumble_job.py
from pyflink.datastream import StreamExecutionEnvironment
from pyflink.table import EnvironmentSettings, StreamTableEnvironment


def create_events_source_kafka(t_env: StreamTableEnvironment) -> str:
    """
    Registers a Kafka source table in the Flink table environment.

    Reads JSON messages from the 'green-trips' topic and exposes
    them as a Flink table with a computed event_timestamp and
    watermark derived from pickup_time.

    Args:
        t_env: The Flink StreamTableEnvironment to register the table in.

    Returns:
        The registered table name ('events').
    """
    table_name = "events"
    source_ddl = f"""
        CREATE TABLE {table_name} (
            PULocationID INTEGER,
            DOLocationID INTEGER,
            trip_distance DOUBLE,
            total_amount DOUBLE,
            tip_amount DOUBLE,
            passenger_count INT,
            pickup_time BIGINT,
            dropoff_time BIGINT,
            event_timestamp AS TO_TIMESTAMP_LTZ(pickup_time, 3),
            WATERMARK for event_timestamp as event_timestamp - INTERVAL '5' SECOND
        ) WITH (
            'connector' = 'kafka',
            'properties.bootstrap.servers' = 'redpanda:29092',
            'topic' = 'green-trips',
            'scan.startup.mode' = 'earliest-offset',
            'properties.auto.offset.reset' = 'earliest',
            'format' = 'json'
        );
        """
    t_env.execute_sql(source_ddl)
    return table_name

def create_events_aggregated_sink(t_env: StreamTableEnvironment) -> str:
    """
    Registers a JDBC sink table in the Flink table environment.

    Creates a PostgreSQL-backed table to store tumbling window
    aggregation results (trip counts and revenue per pickup location).

    Args:
        t_env: The Flink StreamTableEnvironment to register the table in.

    Returns:
        The registered table name ('processed_events_green_flink_5min').
    """

    table_name = 'processed_events_green_flink_5min'
    sink_ddl = f"""
        CREATE TABLE {table_name} (
            window_start TIMESTAMP(3), -- this is created by the tumbling window
            window_end TIMESTAMP(3), -- this is created by the tumbling window
            PULocationID INT,
            num_trips BIGINT,
            total_revenue DOUBLE,
            PRIMARY KEY (window_start, PULocationID) NOT ENFORCED
        ) WITH (
            'connector' = 'jdbc',
            'url' = 'jdbc:postgresql://postgres:5432/postgres',
            'table-name' = '{table_name}',
            'username' = 'postgres',
            'password' = 'postgres',
            'driver' = 'org.postgresql.Driver'
        );
        """
    t_env.execute_sql(sink_ddl)
    return table_name

def log_aggregation():
    """
    Entry point for the Flink streaming job.

    Sets up the execution environment, registers source and sink tables,
    and runs a 5 minute tumbling window aggregation that counts trips and
    sums revenue per PULocationID, writing results to PostgreSQL.
    """

    env = StreamExecutionEnvironment.get_execution_environment()
    env.enable_checkpointing(10 * 1000)
    env.set_parallelism(1)

    settings = EnvironmentSettings \
        .new_instance() \
        .in_streaming_mode() \
        .build()
    
    t_env = StreamTableEnvironment.create(env, environment_settings=settings)

    try:
        source_table = create_events_source_kafka(t_env)
        aggregated_table = create_events_aggregated_sink(t_env)

        t_env.execute_sql(f"""
        INSERT INTO {aggregated_table}
        SELECT
            window_start,
            window_end,
            PULocationID,
            COUNT(*) AS num_trips,
            SUM(total_amount) AS total_revenue
        FROM TABLE(
            TUMBLE(TABLE {source_table}, DESCRIPTOR(event_timestamp), INTERVAL '5' MINUTE)
        )
        GROUP BY window_start, window_end, PULocationID;

        """).wait()

    except Exception as e:
        print("Writing records from Kafka to JDBC failed:", str(e))


if __name__ == '__main__':
    log_aggregation()

